docling_exp/requirements.txt
# FastAPI и сервер
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# Docling
docling==2.63.0

# Конфигурация
pydantic==2.9.2
pyyaml==6.0.2

# Утилиты
aiofiles==24.1.0

# OCR библиотеки с GPU поддержкой
# PyTorch с CUDA поддержкой (для GPU)
# Установка через pip index для CUDA 11.8 или 12.1
# Для CUDA 11.8 используйте: pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118
# Для CUDA 12.1 используйте: pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu121
# Здесь используем стандартную установку, которая автоматически выберет подходящую версию
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# ONNX Runtime с GPU поддержкой (для rapidocr)
# Для GPU версии требуется CUDA и cuDNN
onnxruntime-gpu>=1.16.0

# OCR движки
easyocr>=1.7.0
rapidocr-onnxruntime>=1.3.0
tesserocr>=2.7.0

# Дополнительные зависимости для OCR
opencv-python-headless>=4.8.0
pillow>=10.0.0
numpy>=1.24.0

